requirements.txt
flask
gunicorn
rpi.lgpio
board
adafruit-circuitpython-ssd1306
adafruit-circuitpython-ads1x15
pillow
simplepam